% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{official_palette}
\alias{official_palette}
\title{Official bamOrg Branches' Palettes}
\usage{
official_palette(theme = "marketing")
}
\arguments{
\item{theme}{marketing or finance}
}
\value{
a vector with hex codes
}
\description{
ADD DESCRIPTION
}
